Interface's first usable version
This interface uses pipes in order to communicate with a child program. The child program is the client program, it can either be an executable or a prolog file.
The main module used for creating the interface itself: customtkinter, which is an improvement of tkinter.
Multi-threading was necessary in order to establish a constant bidirectional communication between the parent and the child process. Only the customtkinter object is not threaded, as it should not be.