Small QoL Improvement
Now, whenever the interface is selected/focused in, the input text box is selected for writing.
Also, if the client was not connected to a server the stdin flush in the send_message() method didn't complete properly (it threw an error). This meant that the cleaning of the input text box also didn't occur. Fixing it was done using a simple try except.