Fixed/improved a couple of features
Now the interface is the one that configures the initialization of the server. This will prove useful quite soon, as it will allow the switching of the servers + ports during the session.
Notification sound has been fixed so that it only plays when a message is received (it also used to play for the sending of messages).
The scroll has been fixed so that now you must have the mouse hovering over the text box you wish to scroll.
Last but not least, a problem has been patched. That problem consisted in a pipe error that was thrown because the receive_messages() thread was being initialized before the ctkinter mainloop, as there was no window or anything yet playing for messages to show up.