Improved some features
Created /help command.
Created a special protocol for the message that goes in response to the client.pl's alias question. The client.pl's alias question is for getting an internal alias which is only displayed on the connection and disconnection. This special protocol creates a random number between 0 and 1000000000 and calls it "session_id". This way we are able to distinguish between two different usernames. I made it so that every message includes it.