Huge fix and small Improvement
Now when you answer for the internal alias the external alias (interface's alias) is set to the same value. Further changes with /alias only alter the external alias.
The huge fix refers to a variable which was not being reset and thus made it always use the protocol for answering the alias question.
Also fixed the fact that the ID was not showing for the messages.